* '换肤--黑色' of https://github.com/iunion/YSLive:
  答题器 详情传参 去掉小数
  学生计时器
  学生抢答
  老师抢答器
  计时器
  计时器信令 服务器时间对比
  浮动工具条，背景改透明
  答题器UI
  更改全体静音的图片
  老师答题器
  学生端添加工具箱
  学生工具箱
  YSToolBoxView
  BMImageTitleButtonView

+++ YSLiveManager/YSLiveApiRequest/YSLiveApiRequest.m
@@ -184,7 +184,7 @@
     [parameters bm_setInteger:roomId.integerValue forKey:@"serial"];
     [parameters bm_setString:answerId forKey:@"id"];
     [parameters bm_setDouble:startTime forKey:@"starttime"];
-    [parameters bm_setDouble:endTime forKey:@"endtime"];
+    [parameters bm_setDouble:(int)endTime forKey:@"endtime"];
     [parameters bm_setInteger:0 forKey:@"page"];
     [parameters bm_setInteger:200 forKey:@"pageNum"];
     return [YSApiRequest makeRequestWithURL:urlStr parameters:parameters];
